refactor: 리뷰 후속 유지보수성 정리 3건
- boardService: export * → 외부 소비 타입(BoardViewer/BoardPostDetail/MentionCandidateSource)만 명시 재노출, 내부 헬퍼는 board/ 전용으로 환원(공개 표면 축소)
- settlements: SettlementModals를 features/employer/settlements/로 이동(프레젠테이션 colocate 통일)
- postDetail: BoardDetailComposerContext(Value)를 utils로 이동해 훅→컴포넌트 역방향 의존 제거
- 순수 재배치·재노출, 동작·렌더·시그니처 불변 (tsc 0, test 4286 pass)
